+++ src/app/controllers/app-api.controller.ts
@@ -2,28 +2,35 @@ import { accountApp } from '@/user-account';
 import { ExtendedLogger } from '@/utils';
 import { NextFunction, Request, Response } from 'express';
 import { Express } from 'express-serve-static-core';
+import { NotInitializedException } from '../errors';
 import { authRouter } from '../router/auth.router';
 
-/**
- *
- *
- * @export
- * @class AppAPIController
- */
-export class AppAPIController {
+class AppAPIController {
 	protected logger = new ExtendedLogger(AppAPIController.name);
+	private app: Express = null;
 
-	constructor(private app: Express) {
-		this.logger.debug(`Adding routes to API`);
+	public init(app: Express): void {
+		this.app = app;
 		app.get('/', this.handleRoot());
-		// mount /auth -> passport
-		app.use('/auth', authRouter); // auth related
-		app.use('/account', accountApp); // personal account
+		app.use((req, res, next) => {
+			this.logger.verbose(`API call <${req.method}:${req.url}>`);
+			next();
+		});
+		this.addRouter('/auth', authRouter); // auth related
+		this.addRouter('/account', accountApp); // personal account
 	}
 
-	handleRoot() {
+	public addRouter(mountPoint: string, router: Express): void {
+		if (!this.app) throw new NotInitializedException('Please call init before using this method');
+		this.logger.debug(`mounting router ${mountPoint}`);
+		this.app.use(mountPoint, router);
+	}
+
+	private handleRoot() {
 		return (req: Request, res: Response, next: NextFunction) => {
 			res.redirect('/api-docs');
 		};
 	}
 }
+
+export const APIController = new AppAPIController();