+++ src/app/errors/index.ts
@@ -1,1 +1,2 @@
 export * from './invalid-argument.exception';
+export * from './not-initialized.exception';